Switches SPI and I2C to use enums in the HAL for ports (#531)
Closes #397

+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/hal/I2CJNI.java
@@ -11,15 +11,15 @@ import java.nio.ByteBuffer;
 
 @SuppressWarnings("AbbreviationAsWordInName")
 public class I2CJNI extends JNIWrapper {
-  public static native void i2CInitialize(byte port);
+  public static native void i2CInitialize(int port);
 
-  public static native int i2CTransaction(byte port, byte address, ByteBuffer dataToSend,
+  public static native int i2CTransaction(int port, byte address, ByteBuffer dataToSend,
                                           byte sendSize, ByteBuffer dataReceived, byte receiveSize);
 
-  public static native int i2CWrite(byte port, byte address, ByteBuffer dataToSend, byte sendSize);
+  public static native int i2CWrite(int port, byte address, ByteBuffer dataToSend, byte sendSize);
 
-  public static native int i2CRead(byte port, byte address, ByteBuffer dataRecieved,
+  public static native int i2CRead(int port, byte address, ByteBuffer dataRecieved,
                                    byte receiveSize);
 
-  public static native void i2CClose(byte port);
+  public static native void i2CClose(int port);
 }
