+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/hal/SPIJNI.java
@@ -12,46 +12,46 @@ import java.nio.LongBuffer;
 
 @SuppressWarnings("AbbreviationAsWordInName")
 public class SPIJNI extends JNIWrapper {
-  public static native void spiInitialize(byte port);
+  public static native void spiInitialize(int port);
 
-  public static native int spiTransaction(byte port, ByteBuffer dataToSend,
+  public static native int spiTransaction(int port, ByteBuffer dataToSend,
                                           ByteBuffer dataReceived, byte size);
 
-  public static native int spiWrite(byte port, ByteBuffer dataToSend, byte sendSize);
+  public static native int spiWrite(int port, ByteBuffer dataToSend, byte sendSize);
 
-  public static native int spiRead(byte port, ByteBuffer dataReceived, byte size);
+  public static native int spiRead(int port, ByteBuffer dataReceived, byte size);
 
-  public static native void spiClose(byte port);
+  public static native void spiClose(int port);
 
-  public static native void spiSetSpeed(byte port, int speed);
+  public static native void spiSetSpeed(int port, int speed);
 
-  public static native void spiSetOpts(byte port, int msbFirst, int sampleOnTrailing,
+  public static native void spiSetOpts(int port, int msbFirst, int sampleOnTrailing,
                                        int clkIdleHigh);
 
-  public static native void spiSetChipSelectActiveHigh(byte port);
+  public static native void spiSetChipSelectActiveHigh(int port);
 
-  public static native void spiSetChipSelectActiveLow(byte port);
+  public static native void spiSetChipSelectActiveLow(int port);
 
-  public static native void spiInitAccumulator(byte port, int period, int cmd, byte xferSize,
+  public static native void spiInitAccumulator(int port, int period, int cmd, byte xferSize,
                                                int validMask, int validValue, byte dataShift,
                                                byte dataSize, boolean isSigned, boolean bigEndian);
 
-  public static native void spiFreeAccumulator(byte port);
+  public static native void spiFreeAccumulator(int port);
 
-  public static native void spiResetAccumulator(byte port);
+  public static native void spiResetAccumulator(int port);
 
-  public static native void spiSetAccumulatorCenter(byte port, int center);
+  public static native void spiSetAccumulatorCenter(int port, int center);
 
-  public static native void spiSetAccumulatorDeadband(byte port, int deadband);
+  public static native void spiSetAccumulatorDeadband(int port, int deadband);
 
-  public static native int spiGetAccumulatorLastValue(byte port);
+  public static native int spiGetAccumulatorLastValue(int port);
 
-  public static native long spiGetAccumulatorValue(byte port);
+  public static native long spiGetAccumulatorValue(int port);
 
-  public static native int spiGetAccumulatorCount(byte port);
+  public static native int spiGetAccumulatorCount(int port);
 
-  public static native double spiGetAccumulatorAverage(byte port);
+  public static native double spiGetAccumulatorAverage(int port);
 
-  public static native void spiGetAccumulatorOutput(byte port, LongBuffer value,
+  public static native void spiGetAccumulatorOutput(int port, LongBuffer value,
                                                     LongBuffer count);
 }
